drivers: wifi: morse: add Morse Micro hals and driver
Implement the necessary Morse Micro abstraction layers for Zephyr.
To ensure compatibility with threads on a system not created with
mmosal_create, we will create task metadata on the fly to implement
thread waiters. This utilises the custom data field of the thread
and so this osal is incompatible with other modules relying on the
custom data in a similar way. A magic identifier is used to warn
the developer of such a scenario.

Similarly, a thread abort hook is used to clean up this metadata on
thread termination. Another area of potential symbol collision at link
time.

The Morse Micro Zephyr driver offers a standard Zephyr wifi managment
interface implementing simple ops of scan, connect, disconnect, and status.
Statistics handling is not yet implemented.

A test mode is available which prevents bringup of the netdev. This allows
the host hardware to initialise and enable usage in test applications such
as the Morse Micro porting assistant.