drivers: morse: create task data in task notify
In some conditions, task data may not yet be present as a calling thread
may not have been created with the mmosal api. While logically a wait for
notification is called first, a thread switch could cause the notify to
occur first.

Allow creation of data on tasks which have none.