drivers: morse: get channel from operating element
Legacy versions of mmutils did not support parsing of the operation
element to extract the primary channel of the AP. As this has now
been included in 2.8.2, we use this function instead.

This fixes reporting of the primary channel, which was incorrect.
Previous:
Num  | SSID                             (len) | Chan (Band)   | RSSI | Security             | BSSID             | MFP
1    | MorseMicro                       8     | 50   (UNKNOWN) | -7   | WPA3-SAE-HNP         | 0C:BF:74:00:04:91 | Required
2    | EAP112                           19    | 50   (UNKNOWN) | -43  | WPA3-SAE-HNP         | 28:D0:43:80:F6:6E | Required

This patch:
Num  | SSID                             (len) | Chan (Band)   | RSSI | Security             | BSSID             | MFP
1    | MorseMicro                       8     | 29   (UNKNOWN) | -11  | WPA3-SAE-HNP         | 0C:BF:74:00:04:91 | Required
2    | EAP112                           19    | 33   (UNKNOWN) | -72  | WPA3-SAE-HNP         | 94:BB:43:DC:FA:1C | Required

Note that we report the primary channel, not the operating channel.

Signed-off-by: Arien Judge <arien.judge@morsemicro.com>