morse_sm: Kconfig QoL improvements
On halow_client sample the net_mgmt stack size was being set to 800.
This causes crashes when performing a wifi-scan. Increasing this value
to 2048 resolves this issue and also aligns this driver with the value
in morse_fs.
Additionally increase the TX and RX Pool blocks to 32 to align with
MM-IoT-SDK.

Signed-off-by: Chetan Mistry <chetan.mistry@morsemicro.com>