morse_sm: Implement powersave hooks
The changes adds a rudamentary powersave implementation, tests it on
an mm6108_ekh05_v3 and adds the pm variant.

It can be run by building the alpha_testing sample for the
mm6108_ekh05_v3/stm32u585xx/pm target.
By disabling all of the peripherals and logging we can reduce the power
consumption when running this example down to ~50uA while "idle" (and
connected to an AP).
NB. the "sensors" power-rail must be disabled to see 50uA

Tracked-by: OSP-71
Signed-off-by: Chetan Mistry <chetan.mistry@morsemicro.com>