boards/*/ekh05: Update LED definitions
The EKH05 board has three status LEDs and one RGB tri-LED. These changes
update the device tree to match the board properly.

In detail these changes:
 - Redefine the status leds to the correct pins
 - Add a definition for the RGB led.

The pins for the RGB LED also have signal markings for TIM4_CH1/2/3. As
a result, this change includes a defintion for timers4 and pwm4. NB: it
is not clear if this dts file should define things like the PWM period
or the timer prescaler.

Tracked-by: APP-5875
Signed-off-by: Andrew Dodd <andrew.dodd@morsemicro.com>