boards/*/ekh05: Ensure SPI2 uses DMA
These changes ensure that the SPI2 block (the morse wifi SPI) is hooked
up to use the gpdma1 with the correct settings.

These changes also setup the default for `SPI_STM32_DMA`, as it needs to
be `y` for the hooked up DMA controller to be used by the SPI driver.

The results of this change bump the porting assistant's raw throughput
from 4.0Mbit/s to 6.2Mbit/s.

Tracked-by: APP-5875
Signed-off-by: Andrew Dodd <andrew.dodd@morsemicro.com>