drivers/spi: Update optimised driver to have better performance
The default implementation uses strictly either a DMA transfer or single
byte transfers, depending on whether DMA is enabled. To get better
throughput we need to use single byte transfers when the amount of data
is small, and DMA transfers when it is large.

This change:

 * Updates the driver to include a threshold (default 16) at which point
   is chooses to use a DMA backed transfer instead of single byte
   transfers.

 * Defaults the driver for the EKH05_V3 to use this new `optim` SPI
   driver.

This change improves the raw bit rate seen by the porting assistant from
6 Mbit/s to 10 Mbit/s. Work will continue to improve this (target 20
Mbit/s).

Tracked-by: APP-5875
Signed-off-by: Andrew Dodd <andrew.dodd@morsemicro.com>