boards: mm6108_ekh05_v3: improve power handling for pm variant
To enable low power behaviour with the sensor rail jumper placed,
we want to hold reset lines for BT, SPI Flash and the camera in an
appropriate state. GPIO hogs do this.

Don't disable gpdma, it's used for SPI to the HaLow radio.
Don't disable rng, else we get warnings in the build.
These don't impact power conumption.

Tracked-by: OSP-71
Signed-off-by: Arien Judge <arien.judge@morsemicro.com>