boards: mm8108_ekh05: apply SPI pulldowns for MM8108-EKH05
The MM8108-EKH05 does not have any on-board pullups on the SPI lines
as seen on the MM6108-EKH05. When the host microcontroller enters
deep sleep, this lines will float - and cause a significant increase
in current consumption. Applying the pulldown reduces the idle current
from 500uA to 180uA.

The same pin state needs to be applied to the various GPIOs on board,
including CS, BUSY, and SPI INT, which further reduce consumption to
140uA.

Signed-off-by: Arien Judge <arien.judge@morsemicro.com>